refactor(#9): StudiesContract 잘못된 패키지 변경
